backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
pydantic
requests
python-dotenv
sentence-transformers
hf_xet